refactor: product API

+++ src/product/resolvers/product.resolver.ts
@@ -1,19 +1,17 @@
 import { Resolver, Query, Args, Mutation, Int } from '@nestjs/graphql';
-import { UseGuards } from '@nestjs/common';
-import { JwtAuthGuard, Roles, RolesGuard } from 'src/common';
+import { NotFoundException, UseGuards } from '@nestjs/common';
+import { JwtAuthGuard, Roles, RolesGuard, translationKeys } from 'src/common';
 import { UserRole } from 'src/user';
 import { ProductService } from '../services/product.service';
 import { Product } from '../entities';
 import { CreateProductInput, GetProductsOutput, ProductFilterInput, UpdateProductInput } from '../dtos';
 
+const PRODUCT_NOT_FOUND: string = translationKeys.product.notFound;
 @Resolver(() => Product)
 @UseGuards(JwtAuthGuard, RolesGuard)
 export class ProductResolver {
   constructor(private productService: ProductService) {}
 
-  /**
-   * Retrieves a list of products, with optional filtering, pagination, and limit.
-   */
   @Query(() => GetProductsOutput, { description: 'Retrieve a list of products with optional filters.' })
   async getProducts(
     @Args('filter', { type: () => ProductFilterInput, nullable: true, description: 'Filter options for products.' })
@@ -26,19 +24,19 @@ export class ProductResolver {
     return this.productService.findAllWithFilterAndCount(filter, page, limit);
   }
 
-  /**
-   * Retrieves a single product by its ID.
-   */
   @Query(() => Product, { description: 'Retrieve a single product by its unique identifier.' })
   async getProductById(
     @Args('id', { description: 'The unique identifier of the product.' }) id: string,
   ): Promise<Product> {
-    return this.productService.findOneById(id);
+    const product: Product = await this.productService.findOneById(id);
+
+    if (!product) {
+      throw new NotFoundException(PRODUCT_NOT_FOUND);
+    }
+
+    return product;
   }
 
-  /**
-   * Creates a new product with the given details.
-   */
   @Mutation(() => Product, { description: 'Create a new product.' })
   @Roles(UserRole.admin, UserRole.manager)
   async createProduct(
@@ -48,9 +46,6 @@ export class ProductResolver {
     return this.productService.create(createProductInput);
   }
 
-  /**
-   * Updates an existing product identified by its ID with new details.
-   */
   @Mutation(() => Product, { description: 'Update an existing product.' })
   @Roles(UserRole.admin, UserRole.manager)
   async updateProduct(
@@ -58,17 +53,26 @@ export class ProductResolver {
     @Args('updateProductInput', { description: 'The new details to update the product with.' })
     updateProductInput: UpdateProductInput,
   ): Promise<Product> {
-    return this.productService.findOneAndUpdate(id, updateProductInput);
+    const product: Product = await this.productService.findOneAndUpdate(id, updateProductInput);
+
+    if (!product) {
+      throw new NotFoundException(PRODUCT_NOT_FOUND);
+    }
+
+    return product;
   }
 
-  /**
-   * Deletes a product identified by its ID.
-   */
   @Mutation(() => Boolean, { description: 'Delete a product.' })
   @Roles(UserRole.admin, UserRole.manager)
   async deleteProduct(
     @Args('id', { description: 'The unique identifier of the product to be deleted.' }) id: string,
   ): Promise<boolean> {
+    const product: Product = await this.productService.findOneById(id);
+
+    if (!product) {
+      throw new NotFoundException(PRODUCT_NOT_FOUND);
+    }
+    
     await this.productService.deleteById(id);
     return true;
   }
